New configuration option "domain"
If provided will be used to set document.domain, to allow manipulation
of the walkhub iframe from the root document.

Example of usage: hide the walkhub iframe, define a link with whatever
desing, html/css, in the root document; set onclick in this link to
click a link or button within the walkhub iframe.

But this only works if we can set document.domain

This allows full, simple and flexible control of the design.